Add a dedicated page for sprints/
The default behaviour, namely, that we redirect to the most recent
sprint's details page, is still there. The logic has been moved to a
separate component to simplify things a bit.

Note - the new component doesn't utilize `StateData`. We're about to
ditch it anyway in favor of GraphQL (which does pretty much the same
thing but automagically).