chore(server): add eslint
Added basic eslint configuration based on @hono/eslint-config (hono
is the web framework running under the hood).
I don't know if it's any good, but it's probably better than building
a custom set of rules. And it just works.

Also extended with eslint-prettier-config to silence formatting issues.

Fixed whatever it was complaining about.